refactor: Rename package routingrequest to routerequest
