+++ src/main/java/org/opentripplanner/standalone/config/routerequest/ItineraryFiltersMapper.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.standalone.config.routingrequest;
+package org.opentripplanner.standalone.config.routerequest;
 
 import static org.opentripplanner.standalone.config.framework.json.OtpVersion.NA;
 import static org.opentripplanner.standalone.config.framework.json.OtpVersion.V2_0;
