+++ src/test/java/org/opentripplanner/standalone/config/routerequest/RoutingRequestMapperTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.standalone.config.routingrequest;
+package org.opentripplanner.standalone.config.routerequest;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
